Refactor download routines
The download code was doing the quick part in the routine and the slow
part in the foreground thread.  This is obviously not what we want.
Refactor to do the download of the file in go routines (up to 4 at once)
and then add the name to the channel for processing after.  The wait
group counter is decremented in the process method now instead of in the
download method to get rid of the race condition where we close the
channel before the processing method reads the name from it