Store values and keys in redis
This commit adds the code to store the keys and values in redis.  It
also adds the check to make sure that we do not download an already
processed file.  It does this by storing the archive name in a SET in
redis and checking if the name is present before attempting the download